HACKATHON/backend/requirements.txt
Flask
flask-cors
firebase-admin
google-genai
gspread
google-auth
python-dotenv
flask-limiter
requests
